requirements.txt
# --- FastAPI & server ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Data handling & scientific utils ---
numpy==2.1.1
pandas==2.2.2

# --- ORM & Database (SQLAlchemy + PostgreSQL) ---
SQLAlchemy==2.0.35
psycopg[binary]

# --- Environment management ---
python-dotenv==1.0.1

# --- Pydantic (data validation) ---
pydantic==2.9.2
pydantic-settings==2.4.0
